requirements.txt
Django==5.0.6
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
djangorestframework==3.15.1
Pillow==10.3.0
python-dotenv==1.0.1
py_eureka_client==0.11.9